python-sandbox/extreme-gradient-boosting-with-xgboost/requirements_datacamp.txt
asn1crypto==0.24.0
asttokens==1.1.11
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
cryptography==2.1.4
cycler==0.10.0
decorator==4.4.0
dill==0.2.5
distro-info===0.18ubuntu0.18.04.1
graphviz==0.11
idna==2.8
ipython==4.2.1
ipython-genutils==0.2.0
Jinja2==2.10
joblib==0.13.2
jsonschema==3.0.1
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
markdown2==2.3.7
MarkupSafe==1.1.1
matplotlib==3.1.1
numpy==1.16.4
pandas==0.24.2
pexpect==4.7.0
pickleshare==0.7.5
protobackend==0.2.3
protowhat==1.11.2
ptyprocess==0.6.0
pycrypto==2.6.1
pygobject==3.26.1
pyparsing==2.4.2
pyrsistent==0.15.4
python-apt==1.6.3+ubuntu1
python-dateutil==2.8.0
python-debian==0.1.32
pythonbackend==0.12.0
pythonwhat==2.22.0
pytz==2019.2
pyxdg==0.25
PyYAML==5.1
requests==2.22.0
scikit-learn==0.21.2
scipy==1.3.0
SecretStorage==2.3.1
simplegeneric==0.8.1
six==1.12.0
sklearn-pandas==1.8.0
traitlets==4.3.3
urllib3==1.25.6
xgboost==0.90